requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
greenlet==3.1.1
idna==3.10
importlib_resources==6.4.5
kiwisolver==1.4.7
numpy==2.0.2
packaging==24.1
pillow==11.0.0
playwright==1.48.0
pyee==12.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
